move auth endpoints/logic to blueprint
